+++ pyogrio/tests/test_geopandas_io.py
@@ -3,7 +3,7 @@ import locale
 import os
 import re
 import warnings
-from datetime import datetime
+from datetime import datetime, timedelta, timezone
 from io import BytesIO
 from zipfile import ZipFile
 
@@ -1001,25 +1001,35 @@ def test_write_read_datetime_tz_mixed_offsets(
         assert_geodataframe_equal(result, df)
 
 
+@pytest.mark.parametrize(
+    "dates",
+    [
+        [
+            datetime(2023, 1, 1, 11, 0, 1, 111000),
+            datetime(2023, 6, 1, 10, 0, 1, 111000),
+            np.nan,
+        ],
+        [
+            datetime(2023, 1, 1, 11, 0, 1, 111000, tzinfo=timezone(timedelta(hours=1))),
+            datetime(2023, 6, 1, 10, 0, 1, 111000),
+            np.nan,
+        ],
+    ],
+)
 @pytest.mark.requires_arrow_write_api
-def test_write_read_datetime_tz_mixed_offsets_None(tmp_path, use_arrow):
-    """Test with dates with mixed time zone offsets where some offsets are None.
+def test_write_read_datetime_tz_offsets_None(tmp_path, dates, use_arrow):
+    """Test writing a column with datetimes with and without time zone offsets.
 
-    When datetimes with a mix of having a time zone offsets and some without an offset
-    were written without arrow, this gave the following error:
-        `ValueError: cannot convert float NaN to integer`
+    Two types of errors occured:
+      - For datetimes without any offset, when written without arrow:
+        `TypeError: Invalid dtype float64 for __floordiv__`
+      - For datetimes with a mix of having a time zone offset and without, when written
+        without arrow: `ValueError: cannot convert float NaN to integer`
     """
-    # Pandas datetime64 column types doesn't support mixed time zone offsets, so
-    # it needs to be a list of pandas.Timestamp objects instead.
-    dates = [
-        pd.Timestamp("2023-01-01 11:00:01.111+01:00"),
-        pd.Timestamp("2023-06-01 10:00:01.111+05:00"),
-        pd.Timestamp("2023-06-01 10:00:01.111"),
-        np.nan,
-    ]
-
     df = gp.GeoDataFrame(
-        {"dates": dates, "geometry": [Point(1, 1)] * len(dates)}, crs="EPSG:4326"
+        {"dates": dates, "geometry": [Point(1, 1)] * len(dates)},
+        crs="EPSG:4326",
+        dtype=object,
     )
     fpath = tmp_path / "test.gpkg"
     write_dataframe(df, fpath, use_arrow=use_arrow)
@@ -1030,7 +1040,13 @@ def test_write_read_datetime_tz_mixed_offsets_None(tmp_path, use_arrow):
         mixed_offsets_as_utc=False,
     )
 
-    assert_geodataframe_equal(result, df)
+    exp_df = df.copy()
+    if dates[0].tzinfo is None:
+        exp_df.dates = pd.to_datetime(exp_df.dates, utc=False)
+        if PANDAS_GE_20:
+            exp_df.dates = exp_df.dates.dt.as_unit("ms")
+
+    assert_geodataframe_equal(result, exp_df)
 
 
 @pytest.mark.parametrize("ext", [ext for ext in ALL_EXTS if ext != ".shp"])